# docs/openapi.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "MAIN Smart Parking",
    "version": "1.0.0",
    "description": "SRS 2.3. ADMIN includes USER privileges; OWNER means resource ownership or ADMIN. Business ALLOW/DENY both return 200. Timestamps UTC; history ranges [from,to)."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/device/heartbeats": {
      "post": {
        "operationId": "post__device_heartbeats",
        "tags": [
          "DEVICE"
        ],
        "summary": "POST /device/heartbeats",
        "security": [
          {
            "DeviceId": [],
            "DeviceKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeartbeatResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation / signature error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Role / CSRF denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Missing or not owned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Business or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Body too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Registration/login rate limit retry delay, seconds."
              }
            }
          },
          "503": {
            "description": "Retryable infrastructure failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatRequest"
              }
            }
          }
        }
      }
    },
    "/device/access-events": {
      "post": {
        "operationId": "post__device_access-events",
        "tags": [
          "DEVICE"
        ],
        "summary": "POST /device/access-events",
        "security": [
          {
            "DeviceId": [],
            "DeviceKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation / signature error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Role / CSRF denied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Missing or not owned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Business or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Body too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer"
                },
                "description": "Registration/login rate limit retry delay, seconds."
              }
            }
          },
          "503": {
            "description": "Retryable infrastructure failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "FieldError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "fieldErrors",
          "traceId"
        ],
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "ALREADY_AUTHENTICATED",
              "ALREADY_INSIDE",
              "AUTH_REQUIRED",
              "BODY_TOO_LARGE",
              "CARD_BLOCKED",
              "CSRF_INVALID",
              "CURRENT_PASSWORD_INVALID",
              "DATABASE_UNAVAILABLE",
              "DEVICE_BOOT_NOT_REGISTERED",
              "DEVICE_DISABLED",
              "DEVICE_RATE_LIMITED",
              "DEVICE_UNAUTHORIZED",
              "ENABLED_CARD_EXISTS",
              "FORBIDDEN",
              "IDEMPOTENCY_CONFLICT",
              "INVALID_CREDENTIALS",
              "INVALID_SESSION_STATE",
              "INVALID_SIGNATURE",
              "LOGIN_RATE_LIMITED",
              "NOT_FOUND",
              "ORDER_IN_PROGRESS",
              "PACKAGE_DISABLED",
              "PAYMENT_CONSISTENCY_ERROR",
              "PAYMENT_MODE_MISMATCH",
              "PAYMENT_PROVIDER_UNAVAILABLE",
              "PROVIDER_MISMATCH",
              "RECONCILE_RATE_LIMITED",
              "REGISTRATION_RATE_LIMITED",
              "SESSION_CAPACITY_CONFLICT",
              "SESSION_INVALID",
              "UID_TAKEN",
              "USERNAME_TAKEN",
              "USER_DISABLED",
              "VALIDATION_ERROR"
            ]
          },
          "message": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          },
          "traceId": {
            "type": "string"
          },
          "orderId": {
            "type": "string",
            "format": "uuid",
            "description": "Existing order for ORDER_IN_PROGRESS; only returned within authorized ownership."
          }
        }
      },
...
      "SlotReport": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "slotId",
          "state"
        ],
        "properties": {
          "slotId": {
            "type": "string",
            "enum": [
              "S1",
              "S2",
              "S3"
            ]
          },
          "state": {
            "type": "string",
            "enum": [
              "FREE",
              "OCCUPIED",
              "UNKNOWN"
            ]
          }
        }
      },
      "HeartbeatRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "bootId",
          "seq",
          "uptimeMs",
          "firmwareVersion",
          "slots"
        ],
        "properties": {
          "bootId": {
            "type": "string",
            "format": "uuid"
          },
          "seq": {
            "type": "integer"
          },
          "uptimeMs": {
            "type": "integer"
          },
          "firmwareVersion": {
            "type": "string"
          },
          "slots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SlotReport"
            },
            "minItems": 3,
            "maxItems": 3
          }
        }
      },
      "HeartbeatResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "accepted",
          "serverTime"
        ],
        "properties": {
          "accepted": {
            "type": "boolean"
          },
          "serverTime": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ScanRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "eventId",
          "bootId",
          "gate",
          "cardUid"
        ],
        "properties": {
          "eventId": {
            "type": "string",
            "format": "uuid"
          },
          "bootId": {
            "type": "string",
            "format": "uuid"
          },
          "gate": {
            "type": "string",
            "enum": [
              "IN",
              "OUT"
            ]
          },
          "cardUid": {
            "type": "string"
          }
        }
      },
      "ScanResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "eventId",
          "gate",
          "decision",
          "reason",
          "sessionId",
          "command",
          "openDurationMs",
          "serverTime",
          "validUntil"
        ],
        "properties": {
          "eventId": {
            "type": "string",
            "format": "uuid"
          },
          "gate": {
            "type": "string",
            "enum": [
              "IN",
              "OUT"
            ]
          },
          "decision": {
            "type": "string",
            "enum": [
              "ALLOW",
              "DENY"
            ]
          },
          "reason": {
            "type": "string",
            "enum": [
              "ENTRY_ALLOWED",
              "EXIT_ALLOWED",
              "UNKNOWN_CARD",
              "CARD_BLOCKED",
              "USER_DISABLED",
              "CARD_EXPIRED",
              "ALREADY_INSIDE",
              "OCCUPANCY_UNKNOWN",
              "PARKING_FULL",
              "NO_ACTIVE_SESSION"
            ]
          },
          "sessionId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "command": {
            "type": "string",
            "enum": [
              "OPEN",
              "NONE"
            ]
          },
          "openDurationMs": {
            "type": "integer"
          },
          "serverTime": {
            "type": "string",
            "format": "date-time"
          },
          "validUntil": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "Session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "JSESSIONID"
      },
      "DeviceId": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Device-Id"
      },
      "DeviceKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Device-Key"
      }
    }
  }
}